requirements.txt
Babel>=1.3
Beaker>=1.5.4
Django>=1.8.2
ExifRead>=2.1.1
Flask>=0.10
Flask-Admin>=1.2.0
Flask-SQLAlchemy>=2.0
Flask-Storage>=0.1.0
Flask-Uploads>=0.1.3
GeoIP-Python>=1.2.8
IPy>=0.81
Jinja2>=2.7.3
Mako>=1.0.0
MarkupSafe>=0.23
Paste>=1.7.5.1
Pillow>=2.6.2
PyGithub>=1.25.2
PyXB>=1.2.3
PyYAML>=3.11
SQLAlchemy>=1.0.6
SSSDConfig>=1.12.4
Tempita>=0.5.1
Twisted>=14.0.2
WTForms>=2.0.2
Werkzeug>=0.10.4
backports.ssl-match-hostname>=3.4.0.2
blivet>=0.61.15
characteristic>=14.3.0
chardet>=2.2.1
click>=4.0
cloudinary>=1.1.1
coverage>=3.7
cssselect>=0.9.1
decorator>=3.4.0
deluge>=1.3.11
devassistant>=0.9.4
facebook-sdk>=0.4.0
feedparser>=5.2.0
flickrapi>=2.0
fros>=1.0
github3.py>=0.9.4
glusterfs-api>=3.5.3
gyp>=0.1
httplib2>=0.9.1
humanize>=0.5
iniparse>=0.4
ipython>=3.2.0
itsdangerous>=0.24
javapackages>=4.1.0
kitchen>=1.1.1
langtable>=0.0.29
lxml>=3.3.6
matplotlib>=1.4.3
mock>=1.0.1
nltk>=3.0.1
nose>=1.3.4
ntplib>=0.3.2
numpy>=1.9.2
oauthlib>=0.7.2
pandas>=0.16.2
pexif>=0.15
pexpect>=3.1
policycoreutils-default-encoding>=0.1
progress>=1.2
pubnub>=3.7.2
pwquality>=1.2.4
pyOpenSSL>=0.13.1
pyasn1>=0.1.7
pyasn1-modules>=0.0.6
pycrypto>=2.6.1
pycryptopp>=0.6.0.1206569328141510525648634803928199668821045408958
pycurl>=7.19.3.1
pygame>=1.9.1release
pygobject>=3.14.0
pygpgme>=0.3
pyinotify>=0.9.5
pykickstart>=1.99.65
pyliblzma>=0.5.3
pyparsing>=2.0.3
pyparted>=3.9.5
pyserial>=2.7
python-augeas>=0.5.0
python-dateutil>=2.4.2
python-dmidecode>=3.10.13
python-firebase>=1.2
python-instagram>=1.3.1
python-libtorrent>=1.0.2
python-meh>=0.32
python-nss>=0.16.0
python-twitter>=2.2
pyttsx>=1.1
pytz>=2012d
pyxattr>=0.5.3
pyxdg>=0.25
pyzmq>=14.7.0
requests>=2.5.3
requests-oauthlib>=0.5.0
requests-toolbelt>=0.4.0
rpm-python>=4.12.0.1
schedule>=0.3.2
scsamba>=1.2.100
scsamba.dbus>=1.2.100
seobject>=0.1
sepolicy>=1.1
service-identity>=14.0.0
setproctitle>=1.1.6
setroubleshoot>=1.1
simplejson>=3.5.3
six>=1.9.0
slip>=0.6.0
slip.dbus>=0.6.0
textblob>=0.9.0
tweepy>=3.3.0
uritemplate.py>=0.3.0
urlgrabber>=3.10.1
urllib3>=1.10.2
virtualenv>=12.0.7
wsgiref>=0.1.2
yum-langpacks>=0.4.5
yum-metadata-parser>=1.1.4
zope.event>=4.0.3
zope.interface>=4.1.1
